+++ examples/volumetric/earth_model.py
@@ -37,10 +37,12 @@ msh.add_scalarbar3d(
     label_size=1.5,
     size=[100, 2200],
 )
-# put scalarbar vertical, tell camera to keep bounds into account
+
+# Put scalarbar vertical, tell camera to keep bounds into account
 # msh.scalarbar.rotate_x(90).rotate_z(60).use_bounds(True)
+
 # OR: use clone2d to create a 2D scalarbar from the 3D one
-msh.scalarbar = msh.scalarbar.clone2d(pos=[0.7, -0.95], scale=0.3)
+msh.scalarbar = msh.scalarbar.clone2d(pos=[0.7,-0.95], size=0.3)
  
 # Create cmap for conductor
 cond = conductor.tomesh().cmap(lut, 'cell_scalars', on='cells')