+++ examples/volumetric/slice_plane2.py
@@ -12,6 +12,11 @@ for i in range(4):
 
 amap = [0, 1, 1, 1, 1]  # hide low value points giving them alpha 0
 mslices = merge(slices) # merge all slices into a single Mesh
-mslices.cmap('hot_r', alpha=amap).lighting('off').add_scalarbar3d()
-mslices.scalarbar = mslices.scalarbar.clone2d(pos="bottom-right", scale=0.12)
+mslices.cmap('hot_r', alpha=amap).lighting('off')
+
+# Add a 3d scalarbar to the scene
+mslices.add_scalarbar3d()
+# make 2d copy of it and move it to the center-right
+mslices.scalarbar = mslices.scalarbar.clone2d("center-right", 0.15)
+
 show(vol, mslices, __doc__, axes=1).close()